+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/l2domain/_show.py
@@ -22,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-02-15-preview",
+        "version": "2024-06-15-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/l2isolationdomains/{}", "2024-02-15-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/l2isolationdomains/{}", "2024-06-15-preview"],
         ]
     }
 
@@ -49,6 +49,9 @@ class Show(AAZCommand):
             help="Name of the L2 Isolation Domain.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z]{1}[a-zA-Z0-9-_]{2,127}$",
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -120,7 +123,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-02-15-preview",
+                    "api-version", "2024-06-15-preview",
                     required=True,
                 ),
             }
@@ -184,10 +187,20 @@ class Show(AAZCommand):
                 serialized_name="configurationState",
                 flags={"read_only": True},
             )
+            properties.extended_vlan = AAZStrType(
+                serialized_name="extendedVlan",
+            )
+            properties.last_operation = AAZObjectType(
+                serialized_name="lastOperation",
+                flags={"read_only": True},
+            )
             properties.mtu = AAZIntType()
             properties.network_fabric_id = AAZStrType(
                 serialized_name="networkFabricId",
                 flags={"required": True},
+            )
+            properties.network_to_network_interconnect_id = AAZStrType(
+                serialized_name="networkToNetworkInterconnectId",
             )
             properties.provisioning_state = AAZStrType(
                 serialized_name="provisioningState",
@@ -196,6 +209,11 @@ class Show(AAZCommand):
             properties.vlan_id = AAZIntType(
                 serialized_name="vlanId",
                 flags={"required": True},
+            )
+
+            last_operation = cls._schema_on_200.properties.last_operation
+            last_operation.details = AAZStrType(
+                flags={"read_only": True},
             )
 
             system_data = cls._schema_on_200.system_data