+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_internetgatewayrule.py
@@ -77,6 +77,16 @@ def step_delete(test, checks=None):
     )
 
 
+def step_update(test, checks=None):
+    """Internet Gateway Rule update operation"""
+    if checks is None:
+        checks = []
+    test.cmd(
+        "az networkfabric internetgatewayrule update --resource-group {rg} --resource-name {name} --tags {tag}",
+        checks=checks,
+    )
+
+
 class GA_InternetGatewayRuleScenarioTest1(ScenarioTest):
     """Internet Gateway Rule Scenario test"""
 
@@ -90,6 +100,7 @@ class GA_InternetGatewayRuleScenarioTest1(ScenarioTest):
                 "ruleProperties": CONFIG.get(
                     "INTERNET_GATEWAY_RULE", "rule_properties"
                 ),
+                "tag": CONFIG.get("INTERNET_GATEWAY_RULE", "tag"),
             }
         )