+++ src/workloads/azext_workloads/aaz/latest/workloads/sap_central_instance/_show.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "workloads sap-central-instance show",
-    is_preview=True,
 )
 class Show(AAZCommand):
     """Show the SAP Central Services Instance resource.
@@ -23,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-10-01-preview",
+        "version": "2024-09-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.workloads/sapvirtualinstances/{}/centralinstances/{}", "2023-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.workloads/sapvirtualinstances/{}/centralinstances/{}", "2024-09-01"],
         ]
     }
 
@@ -70,7 +69,7 @@ class Show(AAZCommand):
 
     def _execute_operations(self):
         self.pre_operations()
-        self.SAPCentralInstancesGet(ctx=self.ctx)()
+        self.SapCentralServerInstancesGet(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -85,7 +84,7 @@ class Show(AAZCommand):
         result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
         return result
 
-    class SAPCentralInstancesGet(AAZHttpOperation):
+    class SapCentralServerInstancesGet(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -137,7 +136,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-10-01-preview",
+                    "api-version", "2024-09-01",
                     required=True,
                 ),
             }
@@ -213,12 +212,10 @@ class Show(AAZCommand):
             )
             properties.kernel_patch = AAZStrType(
                 serialized_name="kernelPatch",
-                nullable=True,
                 flags={"read_only": True},
             )
             properties.kernel_version = AAZStrType(
                 serialized_name="kernelVersion",
-                nullable=True,
                 flags={"read_only": True},
             )
             properties.load_balancer_details = AAZObjectType(
@@ -283,7 +280,6 @@ class Show(AAZCommand):
                 flags={"read_only": True},
             )
             enqueue_server_properties.port = AAZIntType(
-                nullable=True,
                 flags={"read_only": True},
             )
 
@@ -296,7 +292,6 @@ class Show(AAZCommand):
                 flags={"read_only": True},
             )
             gateway_server_properties.port = AAZIntType(
-                nullable=True,
                 flags={"read_only": True},
             )
 
@@ -314,17 +309,14 @@ class Show(AAZCommand):
             )
             message_server_properties.http_port = AAZIntType(
                 serialized_name="httpPort",
-                nullable=True,
                 flags={"read_only": True},
             )
             message_server_properties.https_port = AAZIntType(
                 serialized_name="httpsPort",
-                nullable=True,
                 flags={"read_only": True},
             )
             message_server_properties.internal_ms_port = AAZIntType(
                 serialized_name="internalMsPort",
-                nullable=True,
                 flags={"read_only": True},
             )
             message_server_properties.ip_address = AAZStrType(
@@ -333,7 +325,6 @@ class Show(AAZCommand):
             )
             message_server_properties.ms_port = AAZIntType(
                 serialized_name="msPort",
-                nullable=True,
                 flags={"read_only": True},
             )