+++ src/acrquery/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+1.0.1b2
+++++++
+* Bug fix: Removed help command that overrode default `acr -h` command.
+
 1.0.1
 ++++++
 * Add: New mock files to validate query request.